Move SpotifyCard into Componets
