+++ src/pages/Home.js
@@ -1,6 +1,6 @@
 import { useLocation } from "react-router-dom";
 import TwitterCard from "../TwitterSearchCard";
-import SpotifySearchCard from "../SpotifySearchCard";
+import SpotifySearchCard from "../components/SpotifySearchCard";
 import IMDbCard from "../components/IMDbCard";
 import styles from "./Home.module.css";
